Название задачи: Кэшируем Изменил вывод строки черет String.format Изменил тип данных у температуры на int

+++ src/main/java/cache/WeatherData.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public class WeatherData {
     private String city;
-    private double temperature;
+    private int temperature;
     private int humidity;
 }
